move malloc out of parallel region, log current allocated memory

+++ distwt/mpi/context.hpp
@@ -44,7 +44,7 @@ private:
     double m_start_time;
 
     Traffic m_local_traffic;
-    size_t m_alloc_current, m_alloc_max;
+    int64_t m_alloc_current, m_alloc_max;
 
     void count_traffic_tx(size_t target, size_t bytes);
     void count_traffic_rx(size_t source, size_t bytes);
